Got good results with post-ln

+++ elk/utils/typing.py
@@ -24,4 +24,4 @@ def float32_to_int16(x: torch.Tensor) -> torch.Tensor:
 
 def int16_to_float32(x: torch.Tensor) -> torch.Tensor:
     """Converts int16 to float16, then reinterprets as float32."""
-    return x.view(torch.bfloat16).type(torch.float32)
+    return x.view(torch.float16).type(torch.float32)
